LayoutTests/http/tests/security/mixedContent/about-blank-iframe-in-main-frame.html

<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function (e) {
    window.data = e.data;
    shouldBe("window.data", "'Not blocked'");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window iframe that loads an about:blank iframe.  We should
*not* trigger a mixed content callback because the about:blank iframes cannot
be corrupted by active network attackers.</p>
<script>
onload = function() {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-about-blank-frame.html");
}
</script>
</body>
</html>
